Move fields to own dir
