+++ src/components/developer-panel/fields/igt-number-field.vue
