+++ src/components/developer-panel/fields/igt-range-field.vue
